+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * No Distraction for StormBringer theme
 * Fix $el.content for WordPress 5.7
+* Allow shop_manager role to access the admin pages
 
 ### 1.4.1: May 04th, 2021
 * New fields for testing: Real Time Attendance, Lessons Date